// src/config/whiteLabel/types.ts
export type WhiteLabelConfig = {
  brandName: string;
  legalBrandName: string;
  legalEntityName: string;
  websiteUrl: string;
  legalLastUpdated: string;
  contact: {
    name: string;
    address: string;
    email: string;
    phone: string;
    whatsapp: string;
  };
  social: {
    facebookName: string;
    instagramName: string;
  };
};